Added controllers to outsource route logic and implement MVC like architecture
